python/requirements.txt
pillow==11.3.0
python-multipart
fastapi